hotfix: 1. Organization 충돌 해결 - Organization의 QueryDSL 레포지토리 생성에    따른 기존 불필요 메서드 삭제 ( 임시 무효화 ) 2. Lecture QueryDSL 레포지토리 분리 	- Bean 인식을 위해 Online과 Offline의 레포지토리 분리
