+++ src/main/java/com/mzcteam01/mzcproject01be/domains/lecture/repository/queryDsl/QOfflineLectureRepository.java
@@ -1,7 +1,6 @@
 package com.mzcteam01.mzcproject01be.domains.lecture.repository.queryDsl;
 
 import com.mzcteam01.mzcproject01be.domains.lecture.entity.OfflineLecture;
-import com.mzcteam01.mzcproject01be.domains.lecture.entity.OnlineLecture;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.ComparableExpressionBase;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -9,18 +8,17 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-import static com.mzcteam01.mzcproject01be.domains.lecture.entity.QOfflineLecture.*;
-import static com.mzcteam01.mzcproject01be.domains.lecture.entity.QOnlineLecture.*;
+import static com.mzcteam01.mzcproject01be.domains.lecture.entity.QOfflineLecture.offlineLecture;
 
 @RequiredArgsConstructor
-public class LectureRepositoryImpl implements LectureRepositoryCustom {
-
+@Repository
+public class QOfflineLectureRepository {
     private final JPAQueryFactory queryFactory;
 
-    @Override
     public Page<OfflineLecture> findOfflineLectures(Integer SearchType, Pageable pageable) {
 
         List<OfflineLecture> offline  = queryFactory
@@ -39,35 +37,12 @@ public class LectureRepositoryImpl implements LectureRepositoryCustom {
 
     }
 
-    @Override
-    public Page<OnlineLecture> findOnlineLectures(Integer SearchType, Pageable pageable) {
-        List<OnlineLecture> online = queryFactory
-                .selectFrom(onlineLecture)
-                .orderBy(getCreatedOrder(SearchType, onlineLecture.createdAt))
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize())
-                .fetch();
-
-        Long total = queryFactory
-                .select(onlineLecture.count())
-                .from(onlineLecture)
-                .fetchOne();
-
-        return new PageImpl<>(online, pageable, total == null ? 0L : total);
-    }
-
-
     private OrderSpecifier<?> getCreatedOrder(
             Integer searchType,
             ComparableExpressionBase<?> createdAt
-            ) {
+    ) {
         return (searchType == null || searchType == 1) ?
-               createdAt.desc() :
+                createdAt.desc() :
                 createdAt.asc();
     }
-
-
-
-
-
 }
